Create book id implicitly
Book id is duplicated in metadata element of OPF and NCX file.
We choose urn:uuid schema to be default ID for the book. This
schema has been chousen because it's treated specially (is used
for fonts encryption). User is free to add any other ID with
add_identifier method.